baseline 2/requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.6.0

# LangChain & LangGraph
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langgraph==0.2.0
langchain-google-genai==2.0.0

# Embeddings & Vector Store
langchain-huggingface==0.1.0
sentence-transformers==2.2.2
langchain-pinecone==0.2.0
pinecone-client==3.0.0

# Search
tavily-python==0.3.0

# Data Processing
pandas==2.1.0
numpy==1.24.0

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0

# ML (optional)
scikit-learn==1.3.0

# UI
streamlit==1.32.0

# Utilities
tenacity==8.2.3
loguru==0.7.2